skeleton/requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.0
constantly==23.10.4
cryptography==43.0.3
daphne==4.1.2
Django==5.1.3
django-adminlte3==0.1.6
django-allauth==65.2.0
django-htmx==1.21.0
django-jazzmin==3.0.1
django-tinymce==4.1.0
djangorestframework==3.15.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
pillow==11.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==24.2.1
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
service-identity==24.2.0
setuptools==75.5.0
shortuuid==1.0.13
sqlparse==0.5.2
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.2.3
whitenoise==6.8.2
zope.interface==7.1.1